start_slackbot/botmodules/fashion_mnist/requirements.txt
numpy
tensorflow

pillow
